% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_table.R
\name{compass_table}
\alias{compass_table}
\title{Write simple compass indicator table}
\usage{
compass_table(
  country,
  operation,
  year,
  population_type,
  population_rms,
  rms_indicator,
  ridl,
  publish
)
}
\arguments{
\item{country}{iso3 code for the country (easier to recall than the M49 used in the API)}

\item{operation}{operation name}

\item{year}{year to use to extract the baseline from Population Statistics}

\item{population_type}{The list of population type for baseline calculation}

\item{population_rms}{The list of population type covered by RMS}

\item{rms_indicator}{list with indicators and their related frame to pull the value}

\item{ridl}{name of ridl data container to push the data to}

\item{publish}{yes / no}
}
\value{
frame with all compass indicators
}
\description{
The function take the list of calculated variables from an RMS
and output an excel document in the same folder with the correct
format for import into COMPASS.

\if{html}{\out{<div class="sourceCode">}}\preformatted{         The function assumes you have already calculated the indicators and
         have a datalist object with all the correct calculation inside
}\if{html}{\out{</div>}}
}
\examples{
# compass <- export_compass_fill( country = "ECU",
#                        operation = "Ecuador ABC",
#                        year = 2022,
#                         population_type = c("REF","ASY", "OIP"),
#                         population_rms = "Refugees and Asylum-seekers",
#                         rms_indicator =    rbind(
#   c("main", "impact2_2",  "2.2 Proportion of PoCs residing in physically safe and
#                            secure settlements with access to basic facilities"),
#   c("main", "impact2_3",  "2.3 Proportion of PoC with access to health services"),
#   c("P2.S3", "impact3_2a", "3.2a Proportion of PoC enrolled in primary education" ),
#   c("P2.S3", "impact3_2b",  "3.2b Proportion of PoC enrolled in secondary education" ),
#   c("main", "impact3_3",  "3.3 Proportion of PoC feeling safe walking alone in their neighborhood (related SDG 16.1.4)." ),
#   c("S2", "outcome1_2", "1.2 Proportion of children under 5 years of age whose births
#                    have been registered with a civil authority. [SDG 16.9.1 - Tier 1]" ),
#   c("S2", "outcome1_3", "1.3 Proportion of PoC with legally recognized identity documents or credentials [GCR 4.2.2]." ),
#   c("main", "outcome4_1", "4.1 Proportion of PoC who know where to access available GBV services." ),
#   c("main", "outcome4_2", "4.2 Proportion of POCs who do not accept violence against women." ),
#   c("main", "outcome8_2", "8.2 Proportion of PoC with primary reliance on clean (cooking) fuels and technology [SDG 7.1.2 Tier 1]" ),
#   c("main", "outcome9_1", "9.1 Proportion of PoCs living in habitable and affordable housing." ),
#   c("main", "outcome9_2", "9.2 Proportion of PoC that have energy to ensure lighting (close to Sphere)." ),
#   c("main","outcome12_1", "12.1 Proportion of PoC using at least basic drinking water services (SDG)." ),
# #  c("main" , "outcome12_2", "12.2 Proportion of PoC with access to a safe household toilet (SDG)." ),
#   c("main", "outcome13_1", "13.1. Proportion of PoC with an account at a bank or other
#                               financial institution or with a mobile-money-service provider [SDG 8.10.2 Tier 1]." ),
#   c("main", "outcome13_2", "13.2. Proportion of PoC who self-report positive changes in their income compared to previous year." ),
#   c("main", "outcome13_3",  "13.3 Proportion of PoC (working age) who are unemployed." ),
#   c("main", "outcome16_1", "16.1. Proportion of PoC with secure tenure rights and/or
#     property rights to housing and/or land [revised SDG indicator 1.4.2]." )#,
#  # c("main", "outcome16_2", "16.2. Proportion of PoC covered by social protection floors/systems [SDG 1.3.1]." )
#      ),
#              ridl = params$ridl,
#              publish = params$publish )
}
