% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsReport-package.R
\docType{package}
\name{rmsReport-package}
\alias{rmsReport}
\alias{rmsReport-package}
\title{rmsReport: Generate RMS Report}
\description{
Use Rmarkdown template to pull the data from RIDL and organise the content for a summary RMS report.
}
\author{
\strong{Maintainer}: Edouard Legoupil \email{legoupil@unhcr.org}

Other contributors:
\itemize{
  \item TUNHCR [copyright holder]
}

}
\keyword{internal}
