% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_rmsreport.R
\name{template_rmsReport}
\alias{template_rmsReport}
\title{Generate an html / pdf country fact sheet}
\usage{
template_rmsReport(year = 2022, country_asylum_iso3c, folder = "Report")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2022)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{folder}{folder within your project where to put the generated report.
Folder will be created if it does not exist}
}
\value{
nothing the file for the report is generated
}
\description{
This functions provides a quick access to a basic Statistical Fact sheet
template that can be customized according to the needs
}
\examples{

# template_rmsReport(year = 2022, country_asylum_iso3c = "USA",   folder = "Report")


}
